pull request handler closed event handled

+++ backend/src/main/java/com/smartmerge/repository/PullRequestRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PullRequestRepository extends JpaRepository<PullRequest, Integer> {
+public interface PullRequestRepository extends JpaRepository<PullRequest, Long> {
 
     @Transactional
     void deleteAllByInstallationId(int id);
